backend/requirements.txt
Flask==3.0.3
pandas==2.3.3
selenium==4.25.0
gspread==6.1.2
google-auth==2.35.0
openpyxl==3.1.5
Flask-Cors-6.0.1